Updataed sax3 to level of 4

+++ software/3/SAX/screen.h
@@ -23,7 +23,7 @@ class screen
     void draw_title_value(String title, char value, int N, int * values ,int sub_value = 0);
     void draw_title_value(String title, String * value, int N, int * values ,int sub_value = 0);
     void draw_single_value(String value);
-    void draw_standby_screen(int & _octave, int & _transpose, byte _global_mode, bool _delta_mode, byte _joystick, byte _joystick_bis, unsigned int tempo);
+    void draw_standby_screen(int8_t & _octave, int8_t & _transpose, byte _global_mode, bool _delta_mode, byte _joystick, byte _joystick_bis, unsigned int tempo);
     void draw_progress_bar(byte value);
     void draw_low_batt();
 
@@ -38,7 +38,7 @@ class screen
     String previous_title = "aaaa";
     bool changed = false;
 
-  int octave = 0,batterie=255;
+    int octave = 0,batterie=255;
     int transpose = 0;
     byte global_mode = 0;
     bool delta_mode = 0;
